+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Saving models only triggers validations once (OSIDB-3108)
 - Update ACLs of linked objects to match collector flaw (OSIDB-3253)
 - Allow start dates to come from multiple sources in SLA (OSIDB-3221)
+- Update public date for collector flaws (OSIDB-3212)
 
 ### Fixed
 - Cannot modify CVE of existing flaws (OSIDB-3102)